clean-up boxbcast

+++ source/boxbcast.h
@@ -10,7 +10,7 @@ extern void load_boxbcastparms(char *name);
 extern void free_boxbcastdesc(void);
 extern void boxbcasttxtimer(void);
 extern void show_bcastactivity(short unr, cbbproc outputproc);
-extern void bccallback(long l);
+extern void bccallback(int32_t l);
 
 #endif /*BOXBCAST_H*/
